+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 -   Fixed faulty synchronization of chapter text when multiple clients are editing a book that led to temporarily missing chapter text.
 -   Fixed jumpy smooth scrolling of caret position.
 -   Place new empty references at top of list.
+-   Gave edition and chapter editors permission to upload images.
 
 # 0.5.63 - 2023-11-11